+++ packages/serverless-plugin/src/constants.js
@@ -5,7 +5,7 @@ export const SUPPORTED_PROVIDERS = ['aws']
 export const SUPPORTED_RUNTIMES = ['nodejs6.10']
 
 export const INCLUDES = [
-  'node_modules/@serverless-chrome/lambda/package.json',
-  'node_modules/@serverless-chrome/lambda/dist/bundle.cjs.js',
-  'node_modules/@serverless-chrome/lambda/dist/headless-chromium',
+  'node_modules/@quicksprout/lambda/package.json',
+  'node_modules/@quicksprout/lambda/dist/bundle.cjs.js',
+  'node_modules/@quicksprout/lambda/dist/headless-chromium',
 ]